+++ app/src/main/java/com/crevado/fr/clientgqlapp_java/util/AplClient.java
@@ -1,0 +1,24 @@
+package com.crevado.fr.clientgqlapp_java.util;
+
+import com.apollographql.apollo.ApolloClient;
+
+import okhttp3.OkHttpClient;
+
+public class AplClient {
+    public static final String BASE_URL = "https://graphql-p534.onrender.com/graphql";
+    public static final String TAG = "CLIENT Androi APP";
+    private static ApolloClient mApolloClient;
+
+    public static ApolloClient getmApolloClient() {
+        OkHttpClient okHttpClient = new OkHttpClient.Builder()
+                .build();
+
+        mApolloClient = ApolloClient.builder()
+                .serverUrl(BASE_URL)
+                .okHttpClient(okHttpClient)
+                .build();
+
+
+        return mApolloClient;
+    }
+}