+++ CHANGELOG.md
@@ -12,3 +12,4 @@ All notable changes to this project will be documented in this file.
 - feat: add apollo service dependency on app level build.gradle file
 - feat: add the downloaded schema.json file
 - feat: Add UsersQuery and auto-generate UsersQuery class with Apollo
+- feat: Configure Apollo Project and Retrieve GraphQL Data